# trawler/src/main/resources/swagger-spec.json
{"openapi":"3.0.0","paths":{"/api/clusters/{clusterId}/images":{"post":{"operationId":"ImageController_createImage","parameters":[{"name":"clusterId","required":true,"in":"path","schema":{"type":"number"}},{"name":"skipImageScan","required":true,"in":"query","schema":{"type":"boolean"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ImageCreateDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ImageDetailsResponseDto"}}}}},"tags":["M9Sweeper"],"security":[{"jwt-auth":[]},{"x-auth-token":[]}]}},"/api/clusters/{clusterId}/image-by-docker-url":{"get":{"operationId":"ImageController_getImageByDockerUrl","parameters":[{"name":"clusterId","required":true,"in":"path","schema":{"type":"number"}},{"name":"dockerImageUrl","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ImageDetailsResponseDto"}}}}},"tags":["M9Sweeper"],"security":[{"jwt-auth":[]},{"x-auth-token":[]}]}},"/api/clusters/{clusterId}/images/{imageId}/trawler/scan/results/per-policy":{"post":{"operationId":"ImageController_saveImageScanResultsPerPolicy","parameters":[{"name":"clusterId","required":true,"in":"path","schema":{"type":"number"}},{"name":"imageId","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ImageTrawlerResultDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ImageScanResultSaveResponse"}}}}},"tags":["M9Sweeper"],"security":[{"jwt-auth":[]},{"x-auth-token":[]}]}},"/api/clusters/{clusterId}/images/{imageId}/trawler/scan/results":{"post":{"operationId":"ImageController_saveImageScanResults","parameters":[{"name":"clusterId","required":true,"in":"path","schema":{"type":"number"}},{"name":"imageId","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TrawlerScanResults"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ImageScanResultSaveResponse"}}}}},"tags":["M9Sweeper"],"security":[{"jwt-auth":[]},{"x-auth-token":[]}]}},"/api/policies/by-cluster/{clusterId}":{"get":{"operationId":"PolicyController_getPoliciesByClusterId","parameters":[{"name":"clusterId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PoliciesByClusterResponse"}}}}},"tags":["M9Sweeper"],"security":[{"jwt-auth":[]},{"x-auth-token":[]}]}},"/api/clusters/{clusterId}":{"get":{"operationId":"ClusterController_getClusterById","parameters":[{"name":"clusterId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClusterResponse"}}}}},"tags":["M9Sweeper"],"security":[{"jwt-auth":[]},{"x-auth-token":[]}]}},"/api/clusters/by-name/{clusterName}":{"get":{"operationId":"ClusterController_getClusterByClusterName","parameters":[{"name":"clusterName","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ClusterResponse"}}}}},"tags":["M9Sweeper"],"security":[{"jwt-auth":[]},{"x-auth-token":[]}]}},"/api/docker-registries":{"get":{"operationId":"DockerRegistriesController_getDockerRegistries","parameters":[{"name":"page","required":false,"in":"query","schema":{"type":"number"}},{"name":"limit","required":false,"in":"query","schema":{"type":"number"}},{"name":"sort-by","required":false,"in":"query","schema":{"type":"string"}},{"name":"sort-direction","required":false,"in":"query","schema":{"type":"string"}},{"name":"login-required","required":false,"in":"query","schema":{"type":"string"}},{"name":"auth-type","required":false,"in":"query","schema":{"type":"string"}},{"name":"url","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DockerRegistriesResponseDto"}}}}},"tags":["M9Sweeper"],"security":[{"jwt-auth":[]},{"x-auth-token":[]}]}}},"info":{"title":"m9sweeper dash","description":"","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"x-auth-token":{"type":"apiKey","in":"header","name":"x-api-key"}},"schemas":{"ImageCreateDto":{"type":"object","properties":{"url":{"type":"string"},"name":{"type":"string"},"tag":{"type":"string"},"runningInCluster":{"type":"boolean"}},"required":["url","name","tag","runningInCluster"]},"ImageDetailsDto":{"type":"object","properties":{"id":{"type":"number"},"url":{"type":"string"},"name":{"type":"string"},"tag":{"type":"string"},"dockerImageId":{"type":"string"},"clusterId":{"type":"number"},"summary":{"type":"string"}},"required":["id","url","name","tag","dockerImageId","clusterId","summary"]},"ImageDetailsResponseDto":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/ImageDetailsDto"}},"required":["success","message","data"]},"ScanImageIssue":{"type":"object","properties":{"scannerId":{"type":"number"},"scannerName":{"type":"string"},"name":{"type":"string"},"type":{"type":"string"},"vulnerabilityDescUrl":{"type":"string"},"severity":{"type":"string"},"description":{"type":"string"},"isCompliant":{"type":"boolean"},"isFixable":{"type":"boolean"},"wasFixed":{"type":"boolean"},"data":{"type":"string"},"complianceReason":{"type":"string"}},"required":["scannerId","scannerName","name","type","vulnerabilityDescUrl","severity","description","isCompliant","isFixable","wasFixed","data","complianceReason"]},"ImageTrawlerResultDto":{"type":"object","properties":{"summary":{"type":"string"},"criticalIssues":{"type":"number"},"majorIssues":{"type":"number"},"mediumIssues":{"type":"number"},"lowIssues":{"type":"number"},"negligibleIssues":{"type":"number"},"encounterError":{"type":"boolean"},"startedAt":{"type":"number"},"finishedAt":{"type":"number"},"policyId":{"type":"number"},"policyStatus":{"type":"boolean"},"imageHash":{"type":"string"},"issues":{"type":"array","items":{"$ref":"#/components/schemas/ScanImageIssue"}}},"required":["summary","criticalIssues","majorIssues","mediumIssues","lowIssues","negligibleIssues","encounterError","startedAt","finishedAt","policyId","policyStatus","imageHash","issues"]},"ImageScanResultDto":{"type":"object","properties":{"complaint":{"type":"boolean"}},"required":["complaint"]},"ImageScanResultSaveResponse":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/ImageScanResultDto"}},"required":["success","message","data"]},"TrawlerScanResults":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/ImageTrawlerResultDto"}}},"required":["data"]},"ScannerDto":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"type":{"type":"string"},"enabled":{"type":"boolean"},"required":{"type":"boolean"}},"required":["id","name","type","enabled","required"]},"PolicyWithScannerDto":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"enabled":{"type":"boolean"},"enforcement":{"type":"boolean"},"scanners":{"type":"array","items":{"$ref":"#/components/schemas/ScannerDto"}}},"required":["id","name","enabled","enforcement","scanners"]},"PoliciesByClusterResponse":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/PolicyWithScannerDto"}}},"required":["success","message","data"]},"ClusterDto":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"ipAddress":{"type":"string"},"port":{"type":"string"},"context":{"type":"string"},"groupId":{"type":"number"},"tags":{"type":"array","items":{"type":"string"}},"kubeConfig":{"type":"string"},"isEnforcementEnabled":{"type":"boolean"},"isImageScanningEnforcementEnabled":{"type":"boolean"}},"required":["id","name","ipAddress","port","context","groupId","tags","kubeConfig","isEnforcementEnabled","isImageScanningEnforcementEnabled"]},"ClusterResponse":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/ClusterDto"}},"required":["success","message","data"]},"DockerRegistriesDto":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"hostname":{"type":"string"},"loginRequired":{"type":"boolean"},"username":{"type":"string"},"password":{"type":"string"},"authType":{"type":"string"},"authDetails":{"type":"object"},"aliases":{"type":"array","items":{"type":"string"}}},"required":["id","name","hostname","loginRequired","username","password","authType","authDetails","aliases"]},"DockerRegistriesDataDto":{"type":"object","properties":{"totalCount":{"type":"number"},"list":{"type":"array","items":{"$ref":"#/components/schemas/DockerRegistriesDto"}}},"required":["totalCount","list"]},"DockerRegistriesResponseDto":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"$ref":"#/components/schemas/DockerRegistriesDataDto"}},"required":["success","message","data"]}}}}